+++ packages/style-value-types/src/utils.ts
@@ -1,29 +1,5 @@
-// TODO: dedupe with popcorn, atm it causes a dependency cycle though
 export const clamp = (min: number, max: number) => (v: number) =>
   Math.max(Math.min(v, max), min);
-
-/**
- *  Returns a function that will check to see if an argument is
- *  the first characters in the provided `term`
- * `isFirstChars('needle')('haystack')`
- */
-export const isFirstChars = (term: string) => (v: string) =>
-  typeof v === 'string' && v.indexOf(term) === 0;
-
-/*
-  Get value from function string
-  "translateX(20px)" -> "20px"
-*/
-export const getValueFromFunctionString = (value: string) =>
-  value.substring(value.indexOf('(') + 1, value.lastIndexOf(')'));
-
-/*
-  Split comma-delimited string
-
-  "foo,bar" -> ["foo", "bar"]
-*/
-export const splitCommaDelimited = (value: string) =>
-  typeof value === 'string' ? value.split(/,\s*/) : [value];
 
 // If this number is a decimal, make it just five decimal places
 // to avoid exponents
@@ -31,3 +7,4 @@ export const sanitize = (v: number) => (v % 1 ? Number(v.toFixed(5)) : v);
 
 export const floatRegex = /(-)?(\d[\d\.]*)/g;
 export const colorRegex = /(#[0-9a-f]{6}|#[0-9a-f]{3}|#(?:[0-9a-f]{2}){2,4}|(rgb|hsl)a?\((-?\d+%?[,\s]+){2,3}\s*[\d\.]+%?\))/gi;
+export const singleColorRegex = /^(#[0-9a-f]{3}|#(?:[0-9a-f]{2}){2,4}|(rgb|hsl)a?\((-?\d+%?[,\s]+){2,3}\s*[\d\.]+%?\))$/i;